+++ app/Http/Controllers/TourController.php
@@ -143,7 +143,8 @@ class TourController extends Controller
                 \File::delete(public_path('tourimg/' . $filename));
             }
             $filename = date('YmdHis') . $file->getClientOriginalName();
-        } else{
+        } else if(empty($file) && $request->removeImage == 1){
+            $filename = $tour->image_url;
             if (\File::exists(public_path('tourimg/' . $filename))) {
                 \File::delete(public_path('tourimg/' . $filename));
             }
